safari-534.53-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.7_Punctuators/S7.7_A2_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Parse error
S7.7_A2_T3

PASS Expected parsing failure

TEST COMPLETE
